Training of neural network and plot of score histogram for initial features
The file DNN_training.ipynb was developped to train a small neural network, with the features advised by Luca Mollier for the ttm channel. To plot the histogram of the scores, the function plot_hist (previously in feature_analysis.ipynb) was copied in this file and developped to be more general. This function can be put in utils.py in a future commit. There is a problem for the assignement of the test['genWeight'] column to the output_test['genWeight'] column : the non selected rows seem to remain with NaN as values, which causes problems in the plot function. This problem is to be solved.