Update of data_extractor.py and utils.py
Some functions can return several output variables. To treat this case, the output variables can be regrouped in sublists for the assignment, but have to be flattened when put in the output dataframe. To do so, a new function flatten_2D_list was created in utils.py. The data_extractor has not only the structured output_variables as an attribute but also the flattened list. Finally, a new class Data_extractor_test was added to test this new functionality.