Cut on the data loading, functional update and addition of new data_extractors
When the root tree is loaded, a cut is applied to take only events where the sum of the three leptons charges is 1 in absolute value. A condition on the tau jet as well as on the cone around the muons and electrons is also set. The call of the functions is modified : the list of arguments is called with a *. This corresponds to a change of the functions signature in kinematics (and count_tauh in utils) : we have all the variable names instead of *args in the parameter list. A new version Data_extractor_v3 was added, implementing different angles between leptons, as well as invariant masses. Finally, the version Data_extractor_v4 was created, with a pretty complete set of features. At the time of this commit, some of the necessary functions are yet to be coded.