Use of the file model_analysis.ipynb
After retraining, the comparison of accuracies was re-runned. The comparison of the significances was coded and executed, once with all the events mixed up. The image of the result was stored as 'Images\TEST9_global_v4_sig_vs_n_layers.pdf'. For the significance of the separated channels squared summed, I run in the problem of low statistic for one channel : significance computed nowhere -> intesection of cuts is void.