Use of the code for the production of the final plots
The existing files were used to produce the plots for the report. The file dataset_correction.ipynb corrects the weight of the signal events in the TEST9 dataset. The files temp.py and temp2.py produce and store trained DNN for optimal depth selection (careful, these file seem to have a problem, the names of the stored DNN didn't match the DNNS)